+++ .vscode/settings.json
@@ -1,3 +1,3 @@
 {
-    "C_Cpp.errorSquiggles": "Enabled"
+    "C_Cpp.intelliSenseEngineFallback": "Enabled"
 }
