+++ platformio.ini
@@ -12,7 +12,8 @@
 ;platform = espressif32
 ;board = adafruit_feather_esp32_v2
 ;framework = arduino
-[env:main_CMU]
+[env:main]
 platform = atmelavr
-board = uno
+board = nanoatmega328
 framework = arduino
+;upload_port = /dev/cu.wchusbserial14140
