+++ src/main.cpp
@@ -1,7 +1,7 @@
 #include <Arduino.h>
 //#include <Esp.h> // ESP Integration
-#include <SoftwareSerial.h>
-//#include "dpm86CTRL.h"
+//#include <SoftwareSerial.h>
+#include "dpm86CTRL.h"
 
 SoftwareSerial mySerial(2,3); // Rx = 2; Tx = 3
 
@@ -22,35 +22,19 @@ void loop() {
 
   delay(1000);
 
-  Serial.print("Communication(2): ");
-  Serial.print(":01w10=1200,\n"); //command address write volt = 12 fastresponse end
+    Serial.print("Communication(2): ");
+  Serial.print(":01w10=1000.\n"); //command address write volt = 12 fastresponse end
   Serial.println();
   // Serial2
-  mySerial.write(":01w10=1200,\n"); //command address write volt = 12 fastresponse end
+  mySerial.write(":01w10=1000.\n"); //command address write volt = 12 fastresponse end
 
   delay(1000);
 
-  Serial.print("Communication(3): ");
-  Serial.print(":01w10=1200,\r\n"); //command address write volt = 12 fastresponse end
+    Serial.print("Communication(3): ");
+  Serial.print(":01w10=600.\n"); //command address write volt = 12 fastresponse end
   Serial.println();
   // Serial2
-  mySerial.write(":01w10=1200,\r\n"); //command address write volt = 12 fastresponse end
-
-  delay(1000);
-
-  Serial.print("Communication(4): ");
-  Serial.print(":01w10=1200,\r\n"); //command address write volt = 12 fastresponse end
-  Serial.println();
-  // Serial2
-  mySerial.write(':01w10=1200,\r\n'); //command address write volt = 12 fastresponse end
-
-  delay(1000);
-  
-  Serial.print("Communication(5): ");
-  Serial.print(":01w10=1200.\n"); //command address write volt = 12 fastresponse end
-  Serial.println();
-  // Serial2
-  mySerial.print(":01w10=1200.\n"); //command address write volt = 12 fastresponse end
+  mySerial.write(":01w10=600.\n"); //command address write volt = 12 fastresponse end
 
   delay(1000);
 }
